ktext_object: always kzalloc-ate one more byte
This commit ensures that strings on ktext_object_t are always
NULL terminated.
This is not really needed, but bytes are cheap and it's better
to be extra careful with strings.
By design, NULL termination of strings is not required with ktext_object_t
because there is a separate "count" field. But better safe than sorry.
You never know what people are going to do with struct fields.